fix: separated header files and source files
