+++ src/bluetooth/KikurageBLECharacteristicCallbacks.cpp
@@ -1,4 +1,4 @@
-#include "KikurageBLECharacteristicCallbacks.h"
+#include "bluetooth/KikurageBLECharacteristicCallbacks.h"
 
 Color color;
